+++ packages/shared/src/index.ts
@@ -87,3 +87,4 @@ export * from "./socialPlatform.js";
 export * from "./helpdesk.js";
 export * from "./appBuilder.js";
 export * from "./businessIntelligence.js";
+export * from "./enterpriseSearch.js";